Switch to alpine in Docker image
Debian image contains lots of unfixed and its size is huge compare to
alpine image.

The only modification in the code is how we format xfs disks.

For now, we use alpine 3.13 because 3.16 has packages way too recent and
is not supported by the current kernel in most of clusters.