Update logic in Stats tests
Before this commit, the stats were retrieved first using the metrics and
then using the df command. Sometimes, the stats could be reported as
differents because we waited 10 seconds between the two retrievals.

Now, we will try to retrieve the metrics at the same time and we will
try at most 20 times to get an equality.